refactor(fgumi-sort): lift verify_sort_order into fgumi_sort::verify
Move verify_sort_order and VerifySummary from src/lib/commands/sort.rs
into crates/fgumi-sort/src/verify.rs. Relocate the 9 verify-related
tests to crates/fgumi-sort/tests/verify_integration.rs. Wire the new
module into fgumi-sort's public API via pub use. Add fgumi-sort as a
dependency of the main crate so commands/sort.rs can import the lifted
symbols directly.
